+++ src/base/Preprocessing.cpp
@@ -22,7 +22,7 @@ void Preprocessing::apply(Problem& problem, std::vector<double>& x, Multipliers&
          int number_constraints = (int) problem.linear_constraints.size();
          BQPDSolver solver(problem.number_variables, number_constraints, problem.number_variables, true);
 
-         int fortran_indexing = 1;
+         short fortran_indexing = 1;
          CSCMatrix hessian = CSCMatrix::identity(problem.number_variables, fortran_indexing);
          SparseVector linear_objective; // empty
          std::vector<double> d0(problem.number_variables);
